artist-platform/requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
flask-cors==4.0.0
PyJWT==2.8.0
python-multipart==0.0.6
gunicorn==21.2.0
psycopg2-binary==2.9.9
Flask-ReCaptcha==0.4.2
